Fixed spawning and leaderboard.
Also lowered rain box as lag was happening when the rain collided with
the invisible spirit realm (turning the collider off for some reason
reenables the renderer). I also think the rain lifetime should be
lowered and so should the number of particles max as the rain is the
only source of lag atm.